requirements.txt
attrs==22.1.0
certifi==2022.9.14
charset-normalizer==2.1.1
colorama==0.4.5
idna==3.4
iniconfig==1.1.1
mysql-connector-python==8.0.30
packaging==21.3
pluggy==1.0.0
protobuf==3.20.1
py==1.11.0
pyparsing==3.0.9
pytest==7.1.3
requests==2.28.1
tomli==2.0.1
urllib3==1.26.12
